Utils

Description

Module : Utils

Description : Definitions of basic mathematical operations on qubits and gates

Copyright : (c) Mihai Sebastian Ardelean, 2017

License : BSD3

Maintainer : ardeleanasm@gmail.com

Portability : POSIX

Documentation

entangle

Arguments

:: Qubit
first Qubit argument

-> Qubit
second Qubit argument

-> Qubit
return value: Qubit

entangle function is used to perform the Kronecker product between two qubits.

>>> entangle qZero qOne
(4><1)
 [ 0.0 :+ 0.0
 , 1.0 :+ 0.0
 , 0.0 :+ 0.0
 , 0.0 :+ 0.0 ]

apply

Arguments

:: Gate
Gate argument

-> Qubit
Qubit argument

-> Qubit
return value: Qubit

apply function is used to apply a gate on a qubit

>>> apply hGate qZero
 (2><1)
 [ 0.7071067811865475 :+ 0.0
 , 0.7071067811865475 :+ 0.0 ]

(|>)

Arguments

:: Qubit
Qubit argument

-> Gate
Gate argument

-> Qubit
return value: Qubit

|> function have the same effect like apply.

>>> qZero |> hGate
 (2><1)
 [ 0.7071067811865475 :+ 0.0
 , 0.7071067811865475 :+ 0.0 ]

gateProduct

Arguments

:: Gate
Gate argument

-> Gate
Gate argument

-> Gate
return value: Gate

gateProduct function is used to create gates that can be used on multiple qubits

(<+>)

Arguments

:: Gate
Gate argument

-> Gate
Gate argument

-> Gate
return value: Gate

Kronecker product operator is used to create gates that can be used on multiple qubits